uploader/image_processor.py
"""
LiveHub Uploader - Image Processing
"""

from PIL import Image, ImageEnhance


def beautify_image(image_path: str) -> Image.Image:
    """
    Apply image beautification.
    
    Placeholder: Returns original image.
    TODO: Implement actual beautification logic.
    
    Usage example for future implementation:
        enhancer = ImageEnhance.Color(img)
        img = enhancer.enhance(1.1)
        enhancer = ImageEnhance.Contrast(img)
        img = enhancer.enhance(1.05)
    """
    img = Image.open(image_path)
    return img


def load_image(image_path: str) -> Image.Image:
    """Load image from path."""
    return Image.open(image_path)


def resize_image(img: Image.Image, max_size: int = 2048) -> Image.Image:
    """Resize image if too large."""
    if max(img.size) > max_size:
        ratio = max_size / max(img.size)
        new_size = (int(img.size[0] * ratio), int(img.size[1] * ratio))
        return img.resize(new_size, Image.LANCZOS)
    return img
