uploader/requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
watchdog>=3.0.0
requests>=2.31.0
google-auth>=2.22.0
google-auth-oauthlib>=1.1.0
